fix: cleanup component naming
