+++ components/FretboardChart/Fretboard.tsx
