+++ components/FretboardChartLegacy/constants.ts
@@ -1,0 +1,4 @@
+export const NUM_STRINGS = 6;
+export const DEFAULT_HEIGHT = 360;
+export const DEFAULT_FRET_WIDTH = 100;
+export const STROKE = 4;
