--- components/FretboardChartV3/constants.ts
@@ -1,8 +1,0 @@
-export const fontSize = 8;
-export const svgDimension = 100;
-export const fbHeight = 64;
-export const strHeight = fbHeight / 5; // distance between strings
-export const stroke = 1;
-export const circRad = fbHeight / 20;
-export const topSpace = 16;
-export const numStrings = 6;
